Support a local linux development version
